faster reads

+++ max6675.h
@@ -1,5 +1,5 @@
 // this library is public domain. enjoy!
-// www.ladyada.net/learn/sensors/thermocouple
+// https://learn.adafruit.com/thermocouple/
 
 #ifndef ADAFRUIT_MAX6675_H
 #define ADAFRUIT_MAX6675_H
@@ -10,10 +10,11 @@ class MAX6675 {
  public:
   MAX6675(int8_t SCLK, int8_t CS, int8_t MISO);
 
-  double readCelsius(void);
-  double readFahrenheit(void);
+  float readCelsius(void);
+  float readFahrenheit(void);
   // For compatibility with older versions:
-  double readFarenheit(void) { return readFahrenheit(); }
+  float readFarenheit(void) { return readFahrenheit(); }
+
  private:
   int8_t sclk, miso, cs;
   uint8_t spiread(void);
